Update 2.blca_main.py
When blastn finds query matches on the minus strand of the searched sequence, this leads to a malformed -range option when calling blastdbcmd. The -range option of blastdbcmd requires that the start value be less than the end value. In the case where the match is on the (-)strand, blastn outputs the start value as greater than the end value. No error message is generated when this format is passed to blastdbcmd in a batch file, but the -range option is ignored and the match sequence is not trimmed. The result is the entire sequence of the match is written to the .hitdb.fsa file.